auron-rag-master/app/requirements.txt
# FastAPI backend dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.11

# Embeddings
sentence-transformers==5.2.0
torch==2.10.0

# Utilities
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
loguru==0.7.3

# Core dependencies
numpy==2.4.1
requests==2.32.3